+++ src/dev/theskidster/xjge/graphics/Bone.java
@@ -1,6 +1,7 @@
 package dev.theskidster.xjge.graphics;
 
 import org.joml.Matrix4f;
+import org.joml.Vector4f;
 import org.lwjgl.assimp.AIMatrix4x4;
 
 /**